+++ sdk/netapp/azure-resourcemanager-netapp/src/main/java/com/azure/resourcemanager/netapp/fluent/models/BackupProperties.java
@@ -5,86 +5,75 @@
 package com.azure.resourcemanager.netapp.fluent.models;
 
 import com.azure.core.annotation.Fluent;
+import com.azure.core.util.CoreUtils;
 import com.azure.core.util.logging.ClientLogger;
+import com.azure.json.JsonReader;
+import com.azure.json.JsonSerializable;
+import com.azure.json.JsonToken;
+import com.azure.json.JsonWriter;
 import com.azure.resourcemanager.netapp.models.BackupType;
-import com.fasterxml.jackson.annotation.JsonProperty;
+import java.io.IOException;
 import java.time.OffsetDateTime;
 
 /**
  * Backup properties.
  */
 @Fluent
-public final class BackupProperties {
+public final class BackupProperties implements JsonSerializable<BackupProperties> {
     /*
-     * backupId
-     * 
      * UUID v4 used to identify the Backup
      */
-    @JsonProperty(value = "backupId", access = JsonProperty.Access.WRITE_ONLY)
     private String backupId;
 
     /*
-     * creationDate
-     * 
      * The creation date of the backup
      */
-    @JsonProperty(value = "creationDate", access = JsonProperty.Access.WRITE_ONLY)
     private OffsetDateTime creationDate;
 
     /*
      * Azure lifecycle management
      */
-    @JsonProperty(value = "provisioningState", access = JsonProperty.Access.WRITE_ONLY)
     private String provisioningState;
 
     /*
      * Size of backup in bytes
      */
-    @JsonProperty(value = "size", access = JsonProperty.Access.WRITE_ONLY)
     private Long size;
 
     /*
      * Label for backup
      */
-    @JsonProperty(value = "label")
     private String label;
 
     /*
-     * backupType
-     * 
      * Type of backup Manual or Scheduled
      */
-    @JsonProperty(value = "backupType", access = JsonProperty.Access.WRITE_ONLY)
     private BackupType backupType;
 
     /*
      * Failure reason
      */
-    @JsonProperty(value = "failureReason", access = JsonProperty.Access.WRITE_ONLY)
     private String failureReason;
 
     /*
      * ResourceId used to identify the Volume
      */
-    @JsonProperty(value = "volumeResourceId", required = true)
     private String volumeResourceId;
 
     /*
-     * Manual backup an already existing snapshot. This will always be false for scheduled backups and true/false for manual backups
+     * Manual backup an already existing snapshot. This will always be false for scheduled backups and true/false for
+     * manual backups
      */
-    @JsonProperty(value = "useExistingSnapshot")
     private Boolean useExistingSnapshot;
 
     /*
      * The name of the snapshot
      */
-    @JsonProperty(value = "snapshotName")
     private String snapshotName;
 
     /*
      * ResourceId used to identify the backup policy
      */
-    @JsonProperty(value = "backupPolicyResourceId", access = JsonProperty.Access.WRITE_ONLY)
     private String backupPolicyResourceId;
 
     /**
@@ -94,9 +83,7 @@ public final class BackupProperties {
     }
 
     /**
-     * Get the backupId property: backupId
-     * 
-     * UUID v4 used to identify the Backup.
+     * Get the backupId property: UUID v4 used to identify the Backup.
      * 
      * @return the backupId value.
      */
@@ -105,9 +92,7 @@ public final class BackupProperties {
     }
 
     /**
-     * Get the creationDate property: creationDate
-     * 
-     * The creation date of the backup.
+     * Get the creationDate property: The creation date of the backup.
      * 
      * @return the creationDate value.
      */
@@ -154,9 +139,7 @@ public final class BackupProperties {
     }
 
     /**
-     * Get the backupType property: backupType
-     * 
-     * Type of backup Manual or Scheduled.
+     * Get the backupType property: Type of backup Manual or Scheduled.
      * 
      * @return the backupType value.
      */
@@ -258,4 +241,65 @@ public final class BackupProperties {
     }
 
     private static final ClientLogger LOGGER = new ClientLogger(BackupProperties.class);
+
+    /**
+     * {@inheritDoc}
+     */
+    @Override
+    public JsonWriter toJson(JsonWriter jsonWriter) throws IOException {
+        jsonWriter.writeStartObject();
+        jsonWriter.writeStringField("volumeResourceId", this.volumeResourceId);
+        jsonWriter.writeStringField("label", this.label);
+        jsonWriter.writeBooleanField("useExistingSnapshot", this.useExistingSnapshot);
+        jsonWriter.writeStringField("snapshotName", this.snapshotName);
+        return jsonWriter.writeEndObject();
+    }
+
+    /**
+     * Reads an instance of BackupProperties from the JsonReader.
+     * 
+     * @param jsonReader The JsonReader being read.
+     * @return An instance of BackupProperties if the JsonReader was pointing to an instance of it, or null if it was
+     * pointing to JSON null.
+     * @throws IllegalStateException If the deserialized JSON object was missing any required properties.
+     * @throws IOException If an error occurs while reading the BackupProperties.
+     */
+    public static BackupProperties fromJson(JsonReader jsonReader) throws IOException {
+        return jsonReader.readObject(reader -> {
+            BackupProperties deserializedBackupProperties = new BackupProperties();
+            while (reader.nextToken() != JsonToken.END_OBJECT) {
+                String fieldName = reader.getFieldName();
+                reader.nextToken();
+
+                if ("volumeResourceId".equals(fieldName)) {
+                    deserializedBackupProperties.volumeResourceId = reader.getString();
+                } else if ("backupId".equals(fieldName)) {
+                    deserializedBackupProperties.backupId = reader.getString();
+                } else if ("creationDate".equals(fieldName)) {
+                    deserializedBackupProperties.creationDate = reader
+                        .getNullable(nonNullReader -> CoreUtils.parseBestOffsetDateTime(nonNullReader.getString()));
+                } else if ("provisioningState".equals(fieldName)) {
+                    deserializedBackupProperties.provisioningState = reader.getString();
+                } else if ("size".equals(fieldName)) {
+                    deserializedBackupProperties.size = reader.getNullable(JsonReader::getLong);
+                } else if ("label".equals(fieldName)) {
+                    deserializedBackupProperties.label = reader.getString();
+                } else if ("backupType".equals(fieldName)) {
+                    deserializedBackupProperties.backupType = BackupType.fromString(reader.getString());
+                } else if ("failureReason".equals(fieldName)) {
+                    deserializedBackupProperties.failureReason = reader.getString();
+                } else if ("useExistingSnapshot".equals(fieldName)) {
+                    deserializedBackupProperties.useExistingSnapshot = reader.getNullable(JsonReader::getBoolean);
+                } else if ("snapshotName".equals(fieldName)) {
+                    deserializedBackupProperties.snapshotName = reader.getString();
+                } else if ("backupPolicyResourceId".equals(fieldName)) {
+                    deserializedBackupProperties.backupPolicyResourceId = reader.getString();
+                } else {
+                    reader.skipChildren();
+                }
+            }
+
+            return deserializedBackupProperties;
+        });
+    }
 }