+++ sdk/netapp/azure-resourcemanager-netapp/src/main/java/com/azure/resourcemanager/netapp/fluent/models/RegionInfoResourceInner.java
@@ -7,9 +7,12 @@ package com.azure.resourcemanager.netapp.fluent.models;
 import com.azure.core.annotation.Fluent;
 import com.azure.core.management.ProxyResource;
 import com.azure.core.management.SystemData;
+import com.azure.json.JsonReader;
+import com.azure.json.JsonToken;
+import com.azure.json.JsonWriter;
 import com.azure.resourcemanager.netapp.models.RegionInfoAvailabilityZoneMappingsItem;
 import com.azure.resourcemanager.netapp.models.RegionStorageToNetworkProximity;
-import com.fasterxml.jackson.annotation.JsonProperty;
+import java.io.IOException;
 import java.util.List;
 
 /**
@@ -20,14 +23,27 @@ public final class RegionInfoResourceInner extends ProxyResource {
     /*
      * regionInfo properties
      */
-    @JsonProperty(value = "properties")
     private RegionInfoInner innerProperties;
 
     /*
      * Azure Resource Manager metadata containing createdBy and modifiedBy information.
      */
-    @JsonProperty(value = "systemData", access = JsonProperty.Access.WRITE_ONLY)
     private SystemData systemData;
+
+    /*
+     * Fully qualified resource Id for the resource.
+     */
+    private String id;
+
+    /*
+     * The name of the resource.
+     */
+    private String name;
+
+    /*
+     * The type of the resource.
+     */
+    private String type;
 
     /**
      * Creates an instance of RegionInfoResourceInner class.
@@ -54,9 +70,37 @@ public final class RegionInfoResourceInner extends ProxyResource {
     }
 
     /**
-     * Get the storageToNetworkProximity property: Storage to Network Proximity
+     * Get the id property: Fully qualified resource Id for the resource.
      * 
-     * Provides storage to network proximity information in the region.
+     * @return the id value.
+     */
+    @Override
+    public String id() {
+        return this.id;
+    }
+
+    /**
+     * Get the name property: The name of the resource.
+     * 
+     * @return the name value.
+     */
+    @Override
+    public String name() {
+        return this.name;
+    }
+
+    /**
+     * Get the type property: The type of the resource.
+     * 
+     * @return the type value.
+     */
+    @Override
+    public String type() {
+        return this.type;
+    }
+
+    /**
+     * Get the storageToNetworkProximity property: Provides storage to network proximity information in the region.
      * 
      * @return the storageToNetworkProximity value.
      */
@@ -65,9 +109,7 @@ public final class RegionInfoResourceInner extends ProxyResource {
     }
 
     /**
-     * Set the storageToNetworkProximity property: Storage to Network Proximity
-     * 
-     * Provides storage to network proximity information in the region.
+     * Set the storageToNetworkProximity property: Provides storage to network proximity information in the region.
      * 
      * @param storageToNetworkProximity the storageToNetworkProximity value to set.
      * @return the RegionInfoResourceInner object itself.
@@ -82,9 +124,8 @@ public final class RegionInfoResourceInner extends ProxyResource {
     }
 
     /**
-     * Get the availabilityZoneMappings property: Logical availability zone mappings.
-     * 
-     * Provides logical availability zone mappings for the subscription for a region.
+     * Get the availabilityZoneMappings property: Provides logical availability zone mappings for the subscription for a
+     * region.
      * 
      * @return the availabilityZoneMappings value.
      */
@@ -93,9 +134,8 @@ public final class RegionInfoResourceInner extends ProxyResource {
     }
 
     /**
-     * Set the availabilityZoneMappings property: Logical availability zone mappings.
-     * 
-     * Provides logical availability zone mappings for the subscription for a region.
+     * Set the availabilityZoneMappings property: Provides logical availability zone mappings for the subscription for a
+     * region.
      * 
      * @param availabilityZoneMappings the availabilityZoneMappings value to set.
      * @return the RegionInfoResourceInner object itself.
@@ -118,5 +158,50 @@ public final class RegionInfoResourceInner extends ProxyResource {
         if (innerProperties() != null) {
             innerProperties().validate();
         }
+    }
+
+    /**
+     * {@inheritDoc}
+     */
+    @Override
+    public JsonWriter toJson(JsonWriter jsonWriter) throws IOException {
+        jsonWriter.writeStartObject();
+        jsonWriter.writeJsonField("properties", this.innerProperties);
+        return jsonWriter.writeEndObject();
+    }
+
+    /**
+     * Reads an instance of RegionInfoResourceInner from the JsonReader.
+     * 
+     * @param jsonReader The JsonReader being read.
+     * @return An instance of RegionInfoResourceInner if the JsonReader was pointing to an instance of it, or null if it
+     * was pointing to JSON null.
+     * @throws IllegalStateException If the deserialized JSON object was missing any required properties.
+     * @throws IOException If an error occurs while reading the RegionInfoResourceInner.
+     */
+    public static RegionInfoResourceInner fromJson(JsonReader jsonReader) throws IOException {
+        return jsonReader.readObject(reader -> {
+            RegionInfoResourceInner deserializedRegionInfoResourceInner = new RegionInfoResourceInner();
+            while (reader.nextToken() != JsonToken.END_OBJECT) {
+                String fieldName = reader.getFieldName();
+                reader.nextToken();
+
+                if ("id".equals(fieldName)) {
+                    deserializedRegionInfoResourceInner.id = reader.getString();
+                } else if ("name".equals(fieldName)) {
+                    deserializedRegionInfoResourceInner.name = reader.getString();
+                } else if ("type".equals(fieldName)) {
+                    deserializedRegionInfoResourceInner.type = reader.getString();
+                } else if ("properties".equals(fieldName)) {
+                    deserializedRegionInfoResourceInner.innerProperties = RegionInfoInner.fromJson(reader);
+                } else if ("systemData".equals(fieldName)) {
+                    deserializedRegionInfoResourceInner.systemData = SystemData.fromJson(reader);
+                } else {
+                    reader.skipChildren();
+                }
+            }
+
+            return deserializedRegionInfoResourceInner;
+        });
     }
 }