+++ sdk/storage/azure-storage-file-share/src/test/java/com/azure/storage/file/share/ShareApiTests.java
@@ -471,22 +471,6 @@ public class ShareApiTests extends FileShareTestBase {
     }
 
     @Test
-    public void deleteIfExistsDirThatWasAlreadyDeleted() {
-        ShareClient client = premiumFileServiceClient.getShareClient(generateShareName());
-        client.create();
-        Response<Boolean> initialResponse = client.deleteIfExistsWithResponse(null, null, null);
-        // Calling delete again after garbage collection is completed
-        sleepIfRunningAgainstService(45000);
-        Response<Boolean> secondResponse = client.deleteIfExistsWithResponse(null, null, null);
-
-        FileShareTestHelper.assertResponseStatusCode(initialResponse, 202);
-        FileShareTestHelper.assertResponseStatusCode(secondResponse, 404);
-        assertTrue(initialResponse.getValue());
-        assertFalse(secondResponse.getValue());
-    }
-
-
-    @Test
     public void getProperties() {
         primaryShareClient.createWithResponse(testMetadata, 1, null, null);
         Response<ShareProperties> getPropertiesResponse = primaryShareClient.getPropertiesWithResponse(null, null);
@@ -1235,7 +1219,7 @@ public class ShareApiTests extends FileShareTestBase {
 
         ShareStorageException e = assertThrows(ShareStorageException.class, () ->
             aadShareClient.createPermission(permission));
-        assertEquals(ShareErrorCode.AUTHENTICATION_FAILED, e.getErrorCode());
+        assertEquals(ShareErrorCode.INVALID_AUTHENTICATION_INFO, e.getErrorCode());
     }
     @Test
     public void audienceFromString() {